Move models and training to domino module
