+++ src/games/domino/players/__init__.py
@@ -2,7 +2,7 @@ from .strategies import PLAYERS, MergeFactory
 from .hands import HANDS, get_hand, hand_out
 from .behaviors import BEHAVIORS
 from .player import BasePlayer
-from .strategies import state_to_list, selector_maker, encoder_generator, rollout_maker, monte_carlo, get_valids_data
+from .strategies import state_to_list, selector_maker, encoder_generator, rollout_maker, monte_carlo, get_valids_data, alpha_zero_net, AZ_STATE_SHAPE
 
 ALL = [*PLAYERS, *BEHAVIORS]
