+++ src/games/domino/players/strategies/models/__init__.py
@@ -1,0 +1,2 @@
+from .alpha_zero_model import Net as alpha_zero_net
+from .alpha_zero_model import STATE_SHAPE as AZ_STATE_SHAPE
