+++ src/games/domino/training/trainer.py
