--- src/models/__init__.py
@@ -1,2 +1,0 @@
-from .alpha_zero_model import Net as alpha_zero_net
-from .alpha_zero_model import STATE_SHAPE
